requirements.txt
fastapi
uvicorn[standard]
requests
cryptography
eth_account
pynacl